requirements.txt
discord.py
pycrypto
python-dateutil
